fix views

+++ src/main/java/com/zpi/bmarket/bmarket/controllers/OfferController.java
@@ -57,7 +57,7 @@ public class OfferController {
     public String postAddOffer(@ModelAttribute AddOfferDTO offerDTO, Model model, HttpSession session) {
 
         PostStatus status = PostStatus.ERROR;
-        Long id = ((Long) session.getAttribute("userId")).longValue();
+        Long id = (Long) session.getAttribute("userId");
         User user = userRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("id: " + id));
 
         offerDTO.setStatus(statusRepository.findById(0L));  // TODO: jakie statusy - ustawić że oferta jest aktywna
@@ -71,7 +71,7 @@ public class OfferController {
             status = PostStatus.SUCCESS;
         } catch (Exception e) {
             status = PostStatus.DATABASE_ERROR;
-            logger.log(Level.WARNING,"Database Error",e);
+            logger.log(Level.WARNING, "Database Error", e);
         }
 
         model.addAttribute("status", status);
